refactor ezkl.rs
